+++ shared-resources-mc16/src/main/java/nl/enjarai/shared_resources/mc16/mixin/options/GameOptionsMixin.java
@@ -1,0 +1,9 @@
+package nl.enjarai.shared_resources.mc16.mixin.options;
+
+import net.minecraft.client.option.GameOptions;
+import org.spongepowered.asm.mixin.Mixin;
+
+@Mixin(GameOptions.class)
+public class GameOptionsMixin {
+
+}